Move test for normalise_version to flit_core
